move structs-to-data as sample and remove versioning
