--- exercises/version-workflow/practice/workflow_test.go
@@ -1,20 +1,0 @@
-package loanprocess
-
-import (
-	"testing"
-	// "github.com/stretchr/testify/assert"
-	// "go.temporal.io/sdk/worker"
-)
-
-func TestReplayWorkflowHistoryFromFile(t *testing.T) {
-
-	// TODO Create the Workflow Replayer
-
-	// TODO Register the Workflow Definition with the Replayer
-
-	// TODO Replay the Event History from the JSON file you downloaded
-
-	// TODO Use assert.NoError to verify that replaying the history
-	//      does not return an error
-
-}
