--- exercises/version-workflow/solution/shared.go
@@ -1,10 +1,0 @@
-package loanprocess
-
-const TaskQueueName = "loan-processing-workflow-taskqueue"
-
-type ChargeInput struct {
-	CustomerID      string
-	Amount          int
-	PeriodNumber    int
-	NumberOfPeriods int
-}
