+++ samples/using-structs-for-data/activity.go
