+++ samples/using-structs-for-data/microservice/translation-service.go
