+++ samples/using-structs-for-data/shared.go
