+++ samples/using-structs-for-data/workflow.go
